esteid-browser-plugin: changed from int to size_t to fix boost errors
git-svn-id: https://esteid.googlecode.com/svn/esteid-browser-plugin/trunk@2826 0d7e1ef0-d974-fc83-f0b3-ccc071561e0e

+++ esteidAPI.cpp
@@ -520,7 +520,7 @@ std::string esteidAPI::getSigningCertificate() {
     } catch(...) { return ""; } // This API returns nothing on Error
 }
 
-std::string esteidAPI::getSignedHash(std::string hash, int slot) {
+std::string esteidAPI::getSignedHash(std::string hash, size_t slot) {
     WHITELIST_REQUIRED;
     DEPRECATED_CALL;
